+++ openprocurement/auctions/dgf/adapters.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from openprocurement.auctions.core.adapters import AuctionConfigurator
+
+
+class AuctionDgfConfigurator(AuctionConfigurator):
+    name = 'Auction Dgf Configurator'
